Revert "move/rename positronUtils"
This reverts commit 629d6ceda6552fd2091dbe0a97632cf2006d8e86.
