+++ test/smoke/src/areas/positron/console/consoleInput.test.ts
@@ -6,7 +6,7 @@
 
 import { expect } from '@playwright/test';
 import { Application, PositronPythonFixtures, PositronRFixtures } from '../../../../../automation';
-import { setupAndStartApp } from '../../../test-helpers/test-hooks';
+import { setupAndStartApp } from '../../../positronUtils';
 
 describe('Console Input #web', () => {
 	setupAndStartApp();
