+++ test/smoke/src/areas/positron/quarto/quarto.test.ts
@@ -4,9 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Application } from '../../../../../automation';
-// import { setupAndStartApp } from '../../../positronUtils';
+import { setupAndStartApp } from '../../../positronUtils';
 import { expect } from '@playwright/test';
-import { setupAndStartApp } from '../../../test-helpers/test-hooks';
 const path = require('path');
 const fs = require('fs-extra');
