+++ test/smoke/src/areas/positron/variables/notebookVariables.test.ts
@@ -5,7 +5,7 @@
 
 import { expect } from '@playwright/test';
 import { Application, PositronPythonFixtures, PositronRFixtures } from '../../../../../automation';
-import { setupAndStartApp } from '../../../test-helpers/test-hooks';
+import { setupAndStartApp } from '../../../positronUtils';
 
 describe('Variables Pane - Notebook #pr #web', () => {
 	setupAndStartApp();
