+++ test/smoke/src/positronUtils.ts
@@ -5,10 +5,10 @@
 
 import { join } from 'path';
 import * as path from 'path';
-import { Logger, } from '../../../automation';
-import { installAllHandlers, } from '../utils';
-import { createLogger } from './logger';
-import { OPTS } from './config';
+import { Logger, } from '../../automation';
+import { installAllHandlers, } from './utils';
+import { createLogger } from './test-helpers/logger';
+import { OPTS } from './test-helpers/config';
 
 export const ROOT_PATH = join(__dirname, '..', '..', '..');
 const TEST_DATA_PATH = process.env.TEST_DATA_PATH || 'TEST_DATA_PATH not set';
